requirements.txt
# FrothIQ Control Center — Python dependencies
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.30
alembic>=1.13.0

# Cache / pub-sub
redis>=5.0.4
hiredis>=2.3.2

# HTTP client (for frothiq-core proxy)
httpx>=0.27.0

# WebSocket
websockets>=12.0

# Background tasks
anyio>=4.4.0

# Observability
structlog>=24.2.0
prometheus-client>=0.20.0

# Rate limiting
slowapi>=0.1.9

# Dev / test
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-cov>=5.0.0
httpx>=0.27.0
fakeredis[aioredis]>=2.23.0
